move assert_uint256_le to core math lib
