+++ cairo/tests/src/utils/test_uint256.cairo
@@ -4,7 +4,7 @@ from starkware.cairo.common.cairo_builtins import HashBuiltin
 from starkware.cairo.common.uint256 import Uint256
 from starkware.cairo.common.alloc import alloc
 
-from src.utils.uint256 import uint256_to_uint160, uint256_add, uint256_sub, assert_uint256_le
+from src.utils.uint256 import uint256_to_uint160, uint256_add, uint256_sub
 
 func test__uint256_to_uint160{range_check_ptr}() {
     // Given
@@ -48,17 +48,4 @@ func test__uint256_sub{range_check_ptr}() -> Uint256 {
     let (res) = uint256_sub([cast(a_ptr, Uint256*)], [cast(b_ptr, Uint256*)]);
 
     return res;
-}
-
-func test__assert_uint256_le{range_check_ptr}() {
-    alloc_locals;
-    let (a_ptr) = alloc();
-    let (b_ptr) = alloc();
-    %{
-        segments.write_arg(ids.a_ptr, program_input["a"])
-        segments.write_arg(ids.b_ptr, program_input["b"])
-    %}
-    assert_uint256_le([cast(a_ptr, Uint256*)], [cast(b_ptr, Uint256*)]);
-
-    return ();
 }
