+++ python/cairo-core/tests/test_maths.cairo
@@ -1,4 +1,5 @@
-from cairo_core.maths import scalar_to_epns, sign
+from cairo_core.maths import scalar_to_epns, sign, assert_uint256_le
+from starkware.cairo.common.alloc import alloc
 
 func test__sign{range_check_ptr}() -> felt {
     tempvar value;
@@ -12,4 +13,17 @@ func test__scalar_to_epns{range_check_ptr}() -> (
     tempvar scalar;
     %{ ids.scalar = program_input["scalar"] %}
     return scalar_to_epns(scalar);
+}
+
+func test__assert_uint256_le{range_check_ptr}() {
+    alloc_locals;
+    let (a_ptr) = alloc();
+    let (b_ptr) = alloc();
+    %{
+        segments.write_arg(ids.a_ptr, program_input["a"])
+        segments.write_arg(ids.b_ptr, program_input["b"])
+    %}
+    assert_uint256_le([cast(a_ptr, Uint256*)], [cast(b_ptr, Uint256*)]);
+
+    return ();
 }
