+++ python/cairo-core/tests/test_maths.py
@@ -1,5 +1,7 @@
 import pytest
-from hypothesis import assume, given
+from hypothesis import assume, given, settings
+from hypothesis.strategies import integers
+from src.utils.uint256 import int_to_uint256
 from starkware.cairo.lang.cairo_constants import DEFAULT_PRIME
 
 from cairo_addons.testing.strategies import felt, uint128
@@ -28,3 +30,22 @@ class TestMaths:
             assert (
                 sum_p * p_sign - sum_n * n_sign
             ) % DEFAULT_PRIME == scalar % DEFAULT_PRIME
+
+    class TestAssertUint256Le:
+        @given(
+            a=integers(min_value=0, max_value=2**256 - 1),
+            b=integers(min_value=0, max_value=2**256 - 1),
+        )
+        @settings(max_examples=50)
+        def test_assert_uint256_le(self, cairo_run, a, b):
+            if a > b:
+                with pytest.raises(Exception):
+                    cairo_run(
+                        "test__assert_uint256_le",
+                        a=int_to_uint256(a),
+                        b=int_to_uint256(b),
+                    )
+            else:
+                cairo_run(
+                    "test__assert_uint256_le", a=int_to_uint256(a), b=int_to_uint256(b)
+                )
